% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{explore_news}
\alias{explore_news}
\title{Naver News Explorer}
\usage{
explore_news()
}
\value{
없음
}
\description{
네이버 뉴스 탐색 작업을 위한 Shiny 앱 호출
}
\examples{
\dontrun{
 library(koscrap)

 ## 텍스트 데이터 탐색기(Shiny Web Application) 호출
 explore_news()
}
}
\seealso{
\code{\link{search_naver}}
}
\author{
유충현
}
