% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_before_bracket}
\alias{extract_before_bracket}
\title{괄호 앞의 문자열 추출}
\usage{
extract_before_bracket(text)
}
\arguments{
\item{text}{character. 괄호 앞의 문자열을 추출할 텍스트.}
}
\description{
시작 문자부터 열린 괄호 앞까지의 문자열을 추출하는 함수
}
\details{
시작 문자부터 열린 괄호 앞까지의 문자열을 추출하는 함수로 괄호가 없을 경우에는 문자열 전체 반환
}
\examples{
\donttest{
# 사용 예시
text1 <- "이것은 예시 문장(괄호 안의 내용)입니다."
text2 <- "괄호가 없는 문장입니다."

extract_before_bracket(text1)
extract_before_bracket(text2)
}
}
