% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_between_bracket_and_newline}
\alias{extract_between_bracket_and_newline}
\title{중괄호와 개행 사이 문자열 추출}
\usage{
extract_between_bracket_and_newline(text)
}
\arguments{
\item{text}{character. 중괄호와 개행 사이의 문자열을 추출할 텍스트.}
}
\description{
닫힌 중괄호(]) 부터 개행문자 앞까지의 문자열을 추출하는 함수
}
\details{
네이버 뉴스의 본문을 추출하기 위해서 중괄호(])와 개행문자 사이의 문자열을 추출하는 함수 정의
}
\examples{
\donttest{
# 사용 예시
example_text <- "여기는 [키: 값] 첫 번째 추출할 내용\n다음 줄입니다. [태그] 두 번째 추출할 내용\n마지막 줄 [마크] 세 번째 내용\n"
extract_between_bracket_and_newline(example_text)
}
}
