% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_between_brackets}
\alias{extract_between_brackets}
\title{괄호 안의 문자열 추출}
\usage{
extract_between_brackets(text)
}
\arguments{
\item{text}{character. 괄호 안의 문자열을 추출할 텍스트.}
}
\description{
괄호 안의 문자열을 추출하기
}
\details{
열린 괄호와 닫힌 괄호 사이의 문자열을 추출하는 함수로, 여러 개의 괄호가 있을 경우에도 추출 가능함
}
\examples{
\donttest{
# 사용 예시
text <- "이것은 (예시) 문장입니다. 여기 (또 다른 예시)가 있습니다."
extract_between_brackets(text)
}
}
