% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_news.R
\name{extract_news}
\alias{extract_news}
\title{네이버 뉴스 프린트 페이지 html 추출}
\usage{
extract_news(
  url,
  photo = TRUE,
  remove_selectors = NULL,
  n_chars = 150,
  file_name = "minimal_page.html",
  path = "."
)
}
\arguments{
\item{url}{character. 추출을 원하는 뉴스 URL}

\item{photo}{logical. 이미지 포함 여부.}

\item{remove_selectors}{character. html에서 제거할 요소의 사용자 정의.}

\item{n_chars}{integer. 본문에서 추출할 글자 수.}

\item{file_name}{character. 뉴스 기사를 출력할 html 파일 이름. 기본값은 "minimal_page.html".}

\item{path}{character. 뉴스 기사를 파일로 출력할 경로 이름. 기본값은 현재 작업 경로인 ".".}
}
\description{
저작권을 고려해서 네이버 프린트 페이지의 약식 html을 추출합니다.
}
\details{
매체명, 제목, 입력일자, 본문의 일부, 뉴스의 URL등 네이버 뉴스 본문의 일부를 추출합니다.
}
\examples{
\donttest{
url <- "https://n.news.naver.com/article/print/138/0002197754"
extract_news(url)

}

}
