% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_naver.R
\name{get_naver_trend}
\alias{get_naver_trend}
\title{네이버 통합 검색어 트렌드 검색}
\usage{
get_naver_trend(
  keywords = NULL,
  titles = NULL,
  start_date = NULL,
  end_date = NULL,
  time_unit = c("month", "date", "week"),
  device = c("all", "pc", "mo"),
  ages = c("all", "0∼12", "13∼18", "19∼24", "25∼29", "30∼34", "35∼39",
    "40∼44", "45∼49", "50∼54", "55∼59", "60~"),
  gender = c("all", "m", "f"),
  client_id = Sys.getenv("NAVER_CLIENT_ID"),
  client_secret = Sys.getenv("NAVER_CLIENT_SECRET")
)
}
\arguments{
\item{keywords}{character. 주제어에 해당하는 검색어. 최대 20개의 검색어를 설정할 수 있음.}

\item{titles}{character. 주제어. 검색어 묶음을 대표하는 이름.}

\item{start_date}{character. 조회 기간 시작 날짜(yyyy-mm-dd 형식). 2016년 1월 1일부터 조회할 수 있음.}

\item{end_date}{character. 조회 기간 종료 날짜(yyyy-mm-dd 형식).}

\item{time_unit}{character. 구간 단위. "month"(월), "date"(일), "week"(주) 중 하나.}

\item{device}{character. 범위. 검색 환경에 따른 조건. "all"(전체), "pc"(PC), "mo"(모바일) 중 하나.}

\item{ages}{character. 연령대. 검색 사용자의 연령대에 따른 조건. "all"(전체), "0∼12", "13∼18", "19∼24",
"25∼29", "30∼34", "35∼39", "40∼44", "45∼49", "50∼54", "55∼59", "60~" 중 선택. 복수 선택 가능.}

\item{gender}{character. 성별. 검색 사용자의 성별에 따른 조건. "all"(전체), "m"(남성), "f"(여성) 중 하나.}

\item{client_id}{character. 애플리케이션 등록 시 발급받은 Client ID}

\item{client_secret}{character. 애플리케이션 등록 시 발급받은 Client Secret}
}
\value{
tibble
변수 목록은 다음과 같음.:
\itemize{
\item title : character. 검색어의 타이틀(주제어)
\item keywords : character. 주제어에 해당하는 검색어
\item period : character. 구간별 시작 날짜(yyyy-mm-dd 형식)
\item ratio : numeric. 구간별 검색량의 상대적 비율. 구간별 결과에서 가장 큰 값을 100으로 설정한 상댓값.
}
}
\description{
네이버 통합 검색어 트렌드 검색를 출력해주는 REST API를 호출하여, 통합 검색어 트렌드를 검색합니다.
}
\details{
네이버에서 발급받은 Client ID, Client Secret는 개인이 발급받은 키를 사용하며,
유출되어서는 안됩니다.
}
\examples{
\donttest{
# Your authorized API keys
client_id <- "XXXXXXXXXXXXXXXXXXXXXXX"
client_secret <- "XXXXXXXXX"

get_naver_trend(keywords = c("AI,인공지능", "ChatGPT,챗GPT"),
  titles = c("AI", "ChatGPT"),
  start_date = "2024-01-01",
  end_date = "2025-03-15",
  time_unit = "month",
  device = "pc",
  ages = c("13∼18", "19∼24"),
  gender = "all",
  client_id = client_id,
  client_secret = client_secret
  )

}

}
