% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_news.R
\name{news2pdf_minimal}
\alias{news2pdf_minimal}
\title{네이버 뉴스 미니멀 pdf 파일 출력}
\usage{
news2pdf_minimal(
  url = NULL,
  photo = TRUE,
  remove_selectors = NULL,
  n_chars = 150,
  file_name = "minimal_page.pdf",
  path = "."
)
}
\arguments{
\item{url}{character. 네이버 뉴스의 URL.}

\item{file_name}{character. 뉴스 기사를 출력할 pdf 파일 이름. 기본값은 "minimal_page.pdf".}

\item{path}{character. 뉴스 기사를 파일로 출력할 경로 이름. 기본값은 현재 작업 경로인 ".".}
}
\description{
네이버 통합 뉴스의 기사를 프린트 미리보기 미니멀 형식으로 pdf 파일로 출력하기
}
\details{
네이버 뉴스에 게시된 뉴스에 한정해서 출력함. 사용 환경에 chrome이 설치되어 있어야 동작함.
}
\examples{
\donttest{
url <- "https://n.news.naver.com/mnews/article/018/0005981321?sid=101"
news2pdf_minimal(url)
}

}
