% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_naver.R
\name{newscontents}
\alias{newscontents}
\title{네이버 뉴스 기사 추출}
\usage{
newscontents(url = NULL)
}
\arguments{
\item{url}{character. 네이버 뉴스의 URL.}
}
\value{
tibble.
변수 목록은 다음과 같음.:
\itemize{
\item url : character. 네이버 뉴스의 URL
\item title : character. 기사의 타이틀
\item category : character. 기사의 카테고리
\item media_name : character. 기사의 미디어 매체 이름
\item journalist_name : character. 기자의 이름
\item journalist_email : character. 기자의 이메일 주소
\item publish_date : 기사가 작성된 일시
\item update_date : 기사가 수정된 일시
\item article : character. 기사의 본문
}
}
\description{
네이버 통합 뉴스의 기사의 여러 정보를 추출
}
\details{
네이버 뉴스에 게시된 뉴스에 한정해서 뉴스 관련 정보를 추출함.
}
\examples{
\donttest{
url <- "https://n.news.naver.com/mnews/article/018/0005981321?sid=101"
newscontents(url)
}

}
