% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{regist_api_key}
\alias{regist_api_key}
\title{Regist API key to package file}
\usage{
regist_api_key(
  type = c("pubgov", "naver"),
  api_key = NULL,
  client_id = NULL,
  client_secret = NULL
)
}
\arguments{
\item{type}{character. 등록할 API key의 종류.}

\item{api_key}{character. 등록할 공공데이터 포털 API key.}

\item{client_id}{character. 등록할 NAVER client id.}

\item{client_secret}{character. 등록할 NAVER client secret.}
}
\description{
Open API와 인터페이스하기 위한 API key를 등록합니다.
}
\details{
만약에 개인 컴퓨터가 아닌 여러 사용자가 사용하는 환경에 koscrap 패키지를 설치한 경우라면,
API key의 보안을 위해서 regist_api_key()대신 set_api_key()를 사용하세요.
}
\examples{
\dontrun{
# 실제 사용자가 할당받은 공공데이터 포털 API key를 사용합니다.
# regist_api_key("pubgov", api_key = "XXXXXXXXXXX")
}
}
